requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
openai
faiss-cpu
numpy
pydantic
python-dotenv
aiofiles
PyPDF2
requests
sentence-transformers
torch
